fix(mm): remove assumptions about PTE structure
The `mm_populate_table_pte` function should be architecture-independent
but assumes that a PTE can be directly incremented to increment the
physical address it holds. This is not true for the `fake` architecture
used for unit-tests, which shifts the PA depending on the page table
level, so this assumption can lead to page table corruption.

Change-Id: I152adbd7fc7d48f3d0f3477bd75288f10295ee06
Signed-off-by: Alwin Joshy <joshy@qti.qualcomm.com>