fix: relax receivers offset check
Change the check for the receiver permissions structures
(emad descriptor as per the FF-A spec), to ensure it doesn't
overlap with memory region descriptor header instead of forcing
to be placed right after it.

With the above change, the helper 'ffa_memory_region_get_receiver'
needed an update to better detect overflows, due to pointer
arithmetic.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I06892e9ac0ce9b189dc30b4788dacd910acddacd