fix(lifecycle): relax the check for vcpu state in SP_INIT runtime model
A vCPU of SP need not be entirely in STARTING state when it was
allocated CPU cyles in RTM_SP_INIT. For example, this vCPU can
send an direct message to another endpoint and enter BLOCKED state.
Hence, relax the assertion in SP_INIT state machine validation.

For robustness, `vcpu_state_set` utility is leveraged in more vCPU
helpers.

Change-Id: I0e979ea92bbfc542a68f026f12fc31f872d8ce35
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>