fix(ffa_memory): avoid changing receiver's permissions
The memory region descriptor sent by the sender of a memory
sharing operation is cached inside Hafnium's internal state
structures.

The receiver's access permissions entry was getting updated
with those submitted after a successful retrieve request.

This is not necessary as per the current state of the
implementation, and it interferes with the return of
FFA_NS_RES_INFO_GET.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I38c186508d2f3d238c491c411d42ccd7f837dbc9