fix(ffa_memory): retrieve response with correct permissions
Hafnium updated the cached permissions in the memory region
descriptor provided by the sender in a lend/donate/share, with
those requested by the borrower.

This negatively impacted the implementation of FFA_NS_RES_INFO_GET,
given the permissions reported were not those the SP had access to.
Hence producing the fix with the change [1].
[1] However, impacted negatively the permissions used in an Hypervisor
retrieve request, which was tested with TF-A-Tests integration.

This patch refactors and splits the `ffa_retrieved_memory_region_init`
in two new implementations, to handle specifically the response to the
two retrieve requests:
- Partition retrieve request there is only one receiver specified
in the response, and populates the permissions according to what
requested by the borrower in the retrieve request.
- Hypervisor retrieve response there as many receives as specified
by the sender, and populated permissions are those provided
by the sender in the cached memory region descriptor.

Splitting the function helps to keep things simple given the nuances
of both cases, and the dependency to pointer arithmetic in the function.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/43728

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I0c679f6f97166912d13217a37cd548725c6a1338