feat(ff-a): support extended partition info descriptor
This patch adds support for the extended partition info descriptor
as described in the FF-A v1.3 ALP2 specification. The new descriptor
is larger due to the fact that additional properties related to live
activation capability and ff-a version are added.

This means the number of descriptors that can be exported by SPMC
to the caller per invocation is much smaller.

Consequently, new helpers are introduced for legacy endpoints, such
as the v1.1 test endpoints, to work with modified descriptors.

Note that the spec expects the endpoints (i.e. caller) to be aware
of the descriptor layout by looking at the descriptor size field and
making the right conversion.

Change-Id: I1a6b079d2069b500e67b305b7f1da4a65ffb4ddf
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>