feat(lfa): expect live state buffer in partition manifest
An SP is expected to use a live state buffer to save and restore
its working state before and after live activation. Hafnium requires
the platform integrator to declare this buffer as a statically
defined carveout in memory region nodes described in the partition
manifest.

This is an implementation defined approach to make the live activation
simpler as it is difficult for SPMC to infer if the buffer overlaps
with working state of the SP.

Change-Id: Ia08de0eff25a9384671e05bd340a06036f537cb3
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>