feat(lfa): support live activation start request message
An LSP managed by SPMD in EL3 sends a live activation start request
message to SPMC. This patch adds the support in hafnium to process
the incoming message by checking if all conditions are met to start
the live activation of a target partition.

It sends a response back to LSP with appropriate status code.

Change-Id: I37a0d51e166a1b5ce69dd63ef0dbf6c60ecc37e6
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>