feat(lfa): receive stop response for live activation
This patch performs robust handling of partition stop response from SP.
If the SP sends a SUCCESS status, SPMC updates the corresponding VM and
vCPU states to reflect the partition is undergoing live activation
and shall be unavailable. Consequently, SPMC sends a framework message
to LSP with SUCCESS status.

If the SP sends an error status, SPMC puts the partition in aborted
state and propagates the error status to LSP.

Change-Id: I5ba8c7decdf6c8a68dea7dab14c962bc88fa53fd
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>