test(lfa): add helper utilities to exercise live activation
This patch adds a new SP command to increment previously shared buffers
and call it from memory sharing tests. It persists the shared buffer
address on retrieve, map it on demand, and increment on explicit
request.
It also registers the SP IRQ handler which will be needed in live
activation test setup.
It also introduced a small helper routine `check_echo` that is of
use in multiple test scenarios.

Change-Id: I07639c6d74c981e685a4c6b7d75e5200fc677edc
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>