refactor(manifest): add services struct to VM and partition_manifest
In later patches VMs will be able to specify different messaging methods
for each UUID they specify. This pairing of UUIDs and messaging methods
is refered to as a service. Therefore introduce a service struct to the
VM and partition_manifest structs in anticipation of this.

As part of these changes simplify the validation of the UUID whilst
handling DIRECT_REQ2. We also update the dir_msg.c helper functions
to reflect the use of services over the previous UUID list.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Icd34b9284f92a6f1ddeaa192c2328f1489193ccc