feat(manifest): add UUID specific messaging methods
Allow multiple messaging methods to be specified. In this case
the messaging method at index i in the list is supported by the
services with the UUID at index i in the UUID list. If only one
messaging method is specified, it applies to all services.
Also add check that messaging method is not zero as SPs should
support at least one of Direct or Indirect messaging and move
the check that the messaging method supplied is valid to parsing.

Add manifest unit tests to check these changes.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Idd6a5e1b9979b8cf0656d2eb55079ef101d73a8b