fix: defer freeing the memory used for old vCPU structures
A peculiar problem occurs with restarting an SP. If we free the
pages allocated for old vCPUs of an SP and later request for memory
to hold new vCPUs, the set of pages allocated are most likely the
same as the old ones.
During vcpu_switch assembler routine, SPMC performs saving of
hardware state to the C data structures of the previously scheduled
( i.e., old) vCPU. This means we are unintentionally overwriting the
contents of the new vCPU data structure.
Subsequently, SPMC attempts to apply the registers of new vCPU to
the hardware registers but by this time, these registers in the C
data structure were already corrupted.
Hence, defer the freeing of the pages allocated for the old vCPU
structures.

Change-Id: Ia2545e13caccbe8834762abdebd2fdefc62b5140
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>