test(ffa): test for multiple messaging methods
Test that the single messaging method defined in the Service2 partition
manifest applies to all UUIDs by sending a indirect request to both UUID
and UUID2. Also test that if Service3 defines UUID 1 to only support
indirect requests and UUID 2 to support both direct and indirect
requests, a direct request to UUID 1 fails but direct requests to UUID 2
or the NULL UUID succeeds. Note UUID 2 is purposefully the UUID to
support direct requests to ensure for a NULL UUID Hafnium searches for
a UUID that supports direct request and doesn't just check the first
one.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I2f0a4d81dccace3304334ce0d76a06d4a42aedf7