chore: set clang-format PointerAlignment to Right
The Google clang-format style uses PointerAlignment=Left, but the
Hafnium codebase tends to use right. This is usually okay in existing
files because DerivePointerAlignment automatically detects which
alignment to use based on the contents of the file, but clang-format can
occasionally choose right alignment, especially on newly created files,
which is annoying to manually fix if noticed later on. This patch
removes any ambiguity and mandates right-aligned pointer definitions
across the project.

Change-Id: Icce089b782ca2277806859f57d9823488ec19fc5
Signed-off-by: Alwin Joshy <joshy@qti.qualcomm.com>