test(manifest): split properties and children in ManifestDtBuilder
To ensure we always meet the condition that all properties precede subnodes,
track a nodes properties and the properties of the children separately.
This is important with the introduction of the services struct in the
partition manifest as it allows us to use services in the
FfaValidManifest and still add additional properties as required for the
specific tests. This will be done in the next patch.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I910e0717b029e6b8c8f1edc0b45b2903563507fa