fix(smmuv3): use correct size for memory alloc
With the switch to the new memory allocator interface, the SMMUv3
driver's queue allocations were updated to no longer use the mpool
interface. However, the sizes provided to the new interface are still
page counts as required by the old interface, which can result in
insufficient memory being allocated for the queues. This commit updates
the allocations to use the size of the queues in bytes.

Change-Id: I7addcafa0f3d9dbfcf30caad7a95c34d15717dc2
Signed-off-by: Jack Suann <jsuann@qti.qualcomm.com>