fix(mm): incorrect level used to infer PTE during walk
While doing a software page table walk, the recursive helper function
mm_get_range_by_mode_level specified incorrect level while trying
to infer the type of the entry and the Output Address from the PTE.
This would lead to misinterpretation of an L0 page table entry as an L0
sub-table.

It should pass `level` instead of `level-1` to the appropriate helper
functions.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I6768697c8e5933592a5d3d1ae58c1a1d385efe4c